trescolori 2.17.25/requirements.txt
requests==2.31.0
pandas==2.2.3
openpyxl==3.1.2
python-dotenv==1.0.0
replicate==0.22.0
--index-url https://pypi.org/simple
--extra-index-url https://pypi.fury.io/arrow-nightlies/
--prefer-binary 